+++ eduVPN.Views/Windows/ConnectWizard.xaml.cs
@@ -146,7 +146,6 @@ namespace eduVPN.Views.Windows
             // Attach to view model events.
             var view_model = (ViewModels.Windows.ConnectWizard)DataContext;
             view_model.RequestInstanceAuthorization += ConnectWizard_RequestInstanceAuthorization;
-            view_model.RequestTwoFactorEnrollment += ConnectWizard_RequestTwoFactorEnrollment;
             view_model.RequestOpenVPNPasswordAuthentication += ConnectWizard_RequestOpenVPNPasswordAuthentication;
             view_model.RequestOpenVPNUsernamePasswordAuthentication += ConnectWizard_RequestOpenVPNUsernamePasswordAuthentication;
             view_model.RequestTwoFactorAuthentication += ConnectWizard_RequestTwoFactorAuthentication;
@@ -446,36 +445,6 @@ namespace eduVPN.Views.Windows
                 e.CallbackURI = view_model.CallbackURI;
 
             _authorization_popups.Remove(query["state"]);
-        }
-
-        /// <summary>
-        /// Pops-up 2-Factor Enrollment prompt.
-        /// </summary>
-        /// <param name="sender"><see cref="eduVPN.ViewModels.Panels.ConnectingSelectPanel"/> requiring enrollment</param>
-        /// <param name="e">Event arguments. This method fills it with user input.</param>
-        /// <remarks>Occurs when 2-Factor Authentication enrollment is requested.</remarks>
-        private void ConnectWizard_RequestTwoFactorEnrollment(object sender, RequestTwoFactorEnrollmentEventArgs e)
-        {
-            var view_model = new ViewModels.Windows.TwoFactorEnrollmentPopup(sender, e);
-
-            // Create a new 2FA enroll pop-up.
-            var popup = new TwoFactorEnrollmentPopup() { Owner = this, DataContext = view_model };
-            popup.Loaded += (object sender_popup, RoutedEventArgs e_popup) =>
-            {
-                // Set initial focus.
-                if (view_model.SelectedMethod == null && view_model.MethodList.Count > 0)
-                {
-                    view_model.SelectedMethod = view_model.MethodList[0];
-                    if (view_model.MethodList.Count > 1)
-                        (popup.FindName("Method") as Control)?.Focus();
-                }
-            };
-
-            // Set the event args to fill with data to be returned to the event sender.
-            ((Button)popup.FindName("OK")).CommandParameter = e;
-
-            // Run the 2FA enrollment pop-up and pass the credentials to be returned to the event sender.
-            popup.ShowDialog();
         }
 
         /// <summary>