+++ eduVPN/ViewModels/Panels/ConnectingSelectPanel.cs
@@ -13,7 +13,6 @@ using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
-using System.Threading.Tasks;
 
 namespace eduVPN.ViewModels.Panels
 {
@@ -178,42 +177,11 @@ namespace eduVPN.ViewModels.Panels
                 {
                     _connect_selected_profile = new DelegateCommand(
                         // execute
-                        async () =>
+                        () =>
                         {
                             Wizard.ChangeTaskCount(+1);
                             try
                             {
-                                if (SelectedProfile.IsTwoFactorAuthentication)
-                                {
-                                    // Selected profile requires 2-Factor Authentication.
-                                    var authenticating_instance = InstanceSource is LocalInstanceSource ? SelectedProfile.Instance : InstanceSource.AuthenticatingInstance;
-
-                                    // Get user info.
-                                    var userinfo_task = new Task<UserInfo>(() => authenticating_instance.GetUserInfo(authenticating_instance, Window.Abort.Token), TaskCreationOptions.LongRunning);
-                                    userinfo_task.Start();
-                                    var user_info = await userinfo_task;
-
-                                    if (!user_info.IsTwoFactorAuthentication ||
-                                        user_info.TwoFactorMethods != TwoFactorAuthenticationMethods.None && (user_info.TwoFactorMethods & SelectedProfile.TwoFactorMethods) == 0)
-                                    {
-                                        // User is not enrolled for 2FA, or is not enrolled using any required 2FA method.
-
-                                        // Get authenticating instance endpoints. (Already cached by GetUserInfo above. Otherwise it would need to be spawned as a background task to avoid deadlock.)
-                                        var api = authenticating_instance.GetEndpoints(Window.Abort.Token);
-
-                                        // Offer user to enroll for 2FA.
-                                        var e = new RequestTwoFactorEnrollmentEventArgs(user_info, authenticating_instance, SelectedProfile);
-                                        Wizard.Profile_RequestTwoFactorEnrollment(this, e);
-                                        if (e.Credentials == null)
-                                            return;
-
-                                        // Enroll.
-                                        var enrollment_task = new Task(() => authenticating_instance.TwoFactorEnroll(authenticating_instance, e.Credentials, Window.Abort.Token), Window.Abort.Token, TaskCreationOptions.LongRunning);
-                                        enrollment_task.Start();
-                                        await enrollment_task;
-                                    }
-                                }
-
                                 // Set connecting instance.
                                 Wizard.InstanceSourceType = InstanceSourceType;
                                 InstanceSource.ConnectingInstance = SelectedProfile.Instance;