+++ eduVPN/ViewModels/VPN/OpenVPNSession.cs
@@ -522,20 +522,6 @@ namespace eduVPN.ViewModels.VPN
                                         {
                                             switch (e.Message)
                                             {
-                                                case "auth-failure": // Client certificate was deleted/revoked on the server side, or the user is disabled.
-                                                    var user_info = AuthenticatingInstance.GetUserInfo(AuthenticatingInstance, _quit.Token);
-                                                    Wizard.Dispatcher.Invoke(DispatcherPriority.Normal, (Action)(() => UserInfo = user_info)); // BTW, update the UserInfo property with fresh data.
-                                                    if (user_info.IsEnabled)
-                                                    {
-                                                        // We read the user is enabled. Try with a new client certificate then.
-                                                        goto case "tls-error";
-                                                    }
-                                                    else
-                                                    {
-                                                        // User is disabled and this is the most likely cause for "auth-failure". Treat this as any other generic reconnect cases.
-                                                        goto default;
-                                                    }
-
                                                 case "connection-reset": // Connection was reset.
                                                     if (_client_certificate.NotAfter <= DateTime.Now)
                                                     {
@@ -544,6 +530,7 @@ namespace eduVPN.ViewModels.VPN
                                                     }
                                                     goto default;
 
+                                                case "auth-failure": // Client certificate was deleted/revoked on the server side, or the user is disabled.
                                                 case "tls-error": // Client certificate is not compliant with this eduVPN instance. Was eduVPN instance reinstalled?
                                                     // Refresh client certificate.
                                                     _client_certificate = ConnectingProfile.Instance.RefreshClientCertificate(AuthenticatingInstance, _quit.Token);